# Conflicts:
#	workshopCode/src/DomainLayer/TradingSystem/Models/ShoppingCart.java

+++ workshopCode/src/DomainLayer/TradingSystem/Models/ShoppingCart.java
@@ -67,17 +67,22 @@ public class ShoppingCart implements Serializable {
         }
     }
 
-    public void addProduct(String product, Store store, int amount) {
+    public void addProduct(String product, Store store, int amount, boolean persist) {
         if (!baskets.containsKey(store)) {
             Basket b = new Basket(store, this);
+
             //create basket in DB
-            PersistenceController.create(b);
+            if (persist) {
+                PersistenceController.create(b);
+            }
             baskets.put(store, b);
         }
         baskets.get(store).addProduct(store.getProductByName(product), amount);
 
         //update DB
-//        PersistenceController.update(baskets.get(store));
+//        if (persist) {
+//            PersistenceController.update(baskets.get(store));
+//        }
     }
 
     public User getUser() {
@@ -115,7 +120,7 @@ public class ShoppingCart implements Serializable {
         return response.toJSONString();
     }
 
-    public String edit(Store store, String product, int amount) {
+    public String edit(Store store, String product, int amount, boolean persist) {
         Basket basket = baskets.get(store);
         List<ProductItem> items = basket.getProductItems();
         JSONObject response = new JSONObject();
@@ -125,19 +130,27 @@ public class ShoppingCart implements Serializable {
                     items.remove(pi);
                     if (items.isEmpty()) {
                         baskets.remove(store);
+
                         //update DB
-                        PersistenceController.delete(pi);
-                        PersistenceController.delete(baskets);
+                        if (persist) {
+                            PersistenceController.delete(pi);
+                            PersistenceController.delete(baskets);
+                        }
                     } else {
-                        PersistenceController.delete(pi);
+                        if (persist) {
+                            PersistenceController.delete(pi);
+                        }
                     }
                     response.put("SUCCESS", "The product has been updated successfully");
                     return response.toJSONString();
                     //return "The product has been updated successfully";
                 } else {
                     pi.setAmount(amount);
+
                     //update DB
-                    PersistenceController.update(pi);
+                    if (persist) {
+                        PersistenceController.update(pi);
+                    }
                     response.put("SUCCESS", "The product has been updated successfully");
                     return response.toJSONString();
                     //return "The product has been updated successfully";
@@ -243,6 +256,12 @@ public class ShoppingCart implements Serializable {
 
     public String getUserName() {
         return this.userName;
+    }
+
+    public void emptyCart() {
+        this.baskets = new ConcurrentHashMap<Store, Basket>();
+        this.isHistory = false;
+        this.cartTotalPrice = 0;
     }
 
     public int getPaymentTransactionId() {
